refactor(hosting-server-settings): move to the right place, rename and a bit clean up
