+++ client/components/scheduled-updates/scheduled-updates-gate/index.tsx
@@ -3,7 +3,7 @@ import { FC, ReactNode, FocusEvent, useState } from 'react';
 import { useIsEligibleForFeature } from 'calypso/blocks/plugins-scheduled-updates/hooks/use-is-eligible-for-feature';
 import Notice from 'calypso/components/notice';
 import NoticeAction from 'calypso/components/notice/notice-action';
-import HostingActivateStatus from 'calypso/my-sites/hosting/hosting-activate-status';
+import HostingActivateStatus from 'calypso/hosting/server-settings/hosting-activate-status';
 import { useDispatch, useSelector } from 'calypso/state';
 import { transferInProgress } from 'calypso/state/automated-transfer/constants';
 import { getAutomatedTransferStatus } from 'calypso/state/automated-transfer/selectors';
