+++ client/hosting/server-settings/hosting-activate.js
