+++ client/hosting/server-settings/phpmyadmin-card/index.js
