+++ client/hosting/server-settings/phpmyadmin-card/restore-db-password.js
