+++ client/hosting/server-settings/phpmyadmin-card/style.scss
