+++ client/hosting/server-settings/test/hosting-activate-status.jsx
